feat: Update namespace for Role and Action models
This commit updates the namespace for the Role and Action models in the `Role.php` and `Action.php` files. The namespace is changed from `App\Models` to `App\Models\pkg_autorisations`. This change ensures that the models are in the correct namespace and aligns with the package's authorization structure.

Recent user commits:
- Merge branch 'labs-web:develop' into pkg_autorisations__v1-cc1-db-q2-Action
- feat: Add 'nom' and 'controller_id' columns to actions table
- feat: Add 'nom' and 'controller_id' columns to actions table
- feat: Add 'nom' and 'controller_id' columns to actions table
- Merge pull request #14 from Safaa1faiz/pkg_realisation_projet__v10-cc1-db-q1-NatureLivrable
- Merge branch 'develop' into pkg_realisation_projet__v10-cc1-db-q1-NatureLivrable
- Merge pull request #26 from husseinbouik/init-app
- add
- Update user role constants to use "apprenant" instead of "membre"
- Merge pull request #1 from Yasmine-daifane/init-app